added readline
