+++ src/services/OutputService/index.js
@@ -1,3 +1,5 @@
+import { rl } from "../../readline/index.js";
+
 class OutputService {
   constructor() {}
   write({ data, type }) {
@@ -10,11 +12,21 @@ class OutputService {
 
     if (Array.isArray(data)) {
       data.forEach((message) => {
-        console.log(message);
+        rl.write(message + "\n");
       });
       return;
     }
 
+    rl.write(data + "\n");
+  }
+
+  log(data) {
+    if (Array.isArray(data)) {
+      data.forEach((message) => {
+        console.log(message);
+      });
+      return;
+    }
     console.log(data);
   }
 }
